% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreCorum.R
\docType{data}
\name{coreCorum}
\alias{coreCorum}
\title{Core complexes from CORUM database.}
\format{
A data frame with 5280 rows and 21 variables:
\describe{
\item{ComplexID}{ID of the complex}
\item{ComplexName}{Name of the complex}
...
}
}
\source{
\url{http://mips.helmholtz-muenchen.de/corum/#download}
}
\usage{
coreCorum
}
\description{
A dataset containing CORUM complexes from 28.11.2022 Corum 4.1 release,
can be looked up using "View(IMPRINTS.CETSA::coreCorum)"
}
\keyword{datasets}
