% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_average.R
\name{imprints_average}
\alias{imprints_average}
\title{imprints_average}
\usage{
imprints_average(data, savefile = TRUE)
}
\arguments{
\item{data}{dataset after calculating the relative protein abundance differences}

\item{savefile}{a logical to tell if you want save the results or not}
}
\value{
a dataframe
}
\description{
Function to calculate the averaged signals from the IMPRINTS-CETSA result
}
\examples{
\dontrun{
  MOLM <- imprints_average(MOLM)
}


}
