% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_barplotting.R
\name{imprints_barplotting}
\alias{imprints_barplotting}
\title{imprints_barplotting}
\usage{
imprints_barplotting(
  data,
  treatmentlevel = NULL,
  setlevel = NULL,
  corrtable = NULL,
  plotseq = NULL,
  printBothName = TRUE,
  printGeneName = FALSE,
  pfdatabase = FALSE,
  witherrorbar = TRUE,
  colorpanel = c("gray", "blue", "orange"),
  usegradient = FALSE,
  colorgradient = c("#4575B4", "ivory", "#D73027"),
  linegraph = FALSE,
  log2scale = TRUE,
  ratio = 0.6,
  layout = NULL,
  ret_plot = FALSE,
  save_pdf = TRUE,
  external = TRUE,
  toplabel = "IMPRINTS-CETSA bar plotting",
  leftlabel = "",
  bottomlabel = "",
  pdfname = "bar_ggplotting.pdf",
  pdfheight = 12,
  pdfwidth = 12
)
}
\arguments{
\item{data}{dataset after imprints_caldiff() to plot}

\item{treatmentlevel}{a vector of treatment labels, such as c("DMSO","TNFa","AT26533")
the order determines the arrangement, so in this case DMSO group would be the first group}

\item{setlevel}{a vector of set information if any, such as c("M13","M16")}

\item{corrtable}{a correlation table, could be useful for customized ranking}

\item{plotseq}{a vector of plots arrangement sequence (in composite ID)}

\item{printBothName}{a logical to tell if you want to print the both protein names on the plot}

\item{printGeneName}{a logical to tell if you want to print the gene names on the plot}

\item{pfdatabase}{a logical for using P.falciparum database or not}

\item{witherrorbar}{a logical to print or not the error bar on the plot}

\item{colorpanel}{a vector of customizable color scheme provided by the user, default set
c("gray","blue","orange")}

\item{usegradient}{whether the barplot should be draw in color gradient format}

\item{colorgradient}{the color scheme of gradient applied, default value c("#4575B4","ivory", "#D73027")}

\item{linegraph}{whether to plot the graph in a line graph format, default set to FALSE}

\item{log2scale}{whether the yscales should be in log2 scale, default set to TRUE}

\item{ratio}{aspect ratio of the plot, default set to 0.6}

\item{layout}{a vector indicating the panel layout for multi-panel plots
per page, default value is c(2,3) for set containing data, otherwise c(4,3)}

\item{ret_plot}{a logical to tell if you want to return the plots object}

\item{save_pdf}{a logical to tell if you want to save plots in a pdf file}

\item{toplabel}{textual label at the top part of the page}

\item{leftlabel}{textual label at the left side of the page}

\item{bottomlabel}{textual label at the bottom part of the page}

\item{pdfheight}{a number indicate the height of pdf file, default value 12}

\item{pdfwidth}{a number indicate the width of pdf file, default value 12}
}
\value{
a list of ggplot2 object
}
\description{
Function to generate pdf files with multipanel bar plots for IMPRINTS-CETSA data
}
\examples{
\dontrun{
  imprints_barplotting(MOLM, treatmentlevel=c("DMSO","TNFa","AT26533"), setlevel=c("M13","M16"))
}


}
