% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_caldiff_f.R
\name{imprints_caldiff_f}
\alias{imprints_caldiff_f}
\title{imprints_caldiff_f}
\usage{
imprints_caldiff_f(data, reftreatment = NULL, withinrep = TRUE)
}
\arguments{
\item{data}{Normalized dataset to calculate the relative protein abundance differences}

\item{reftreatment}{the name of treatment used as the reference control, such as "DMSO" or "Ctrl"}

\item{withinrep}{whether the calculation of the relative protein abundance difference should
still within the same biorep, default set to TRUE, when the bioreps are balanced}
}
\value{
a dataframe
}
\description{
Function to calculate the pair-wise (per replicate and temperature) protein abundance differences,
This one returns the same results as imprints_caldiff(), but in a much faster manner
}
\examples{
\dontrun{
  MOLM <- imprints_caldiff_f(MOLM, reftreatment="DMSO")
}


}
