% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_complex_mapping.R
\name{imprints_complex_mapping}
\alias{imprints_complex_mapping}
\title{imprints_complex_mapping}
\usage{
imprints_complex_mapping(
  data,
  categorytable = NULL,
  set = NULL,
  treatment = NULL,
  targetcategory = c("NC", "CN", "CC"),
  complexdatabase = "Corum",
  organism = "Human",
  complexID = NULL,
  minsubunitsIdentified = 3,
  cutisoformbycount = TRUE
)
}
\arguments{
\item{data}{dataset after calculating the relative protein abundance differences
and deriving the average of the profile, i.e., imprints_average(), make sure the columns
with readings are named in the format like "37C_Treatment"}

\item{categorytable}{dataset after imprints_score() or imprints_globalview(), the data
should contain one column named after "category", to allow focusing on a subset of the data,
say the "Changed" proteins}

\item{set}{a single character to indicate the sample name to analyze if any}

\item{treatment}{a single character to indicate the sample name to analyze}

\item{targetcategory}{the default value is c("NC","CN","CC")}

\item{complexdatabase}{by default is core Corum database, 03.09.2018 Corum 3.0 current release}

\item{organism}{choose from Human/Mouse/Rat, default is Human}

\item{complexID}{a vector of complexID to retrieve from the data, if specified}

\item{minsubunitsIdentified}{the minimal number of subunits should be identified,
default value is 3}

\item{cutisoformbycount}{when there are several isoforms mapped in a complex, by default to
filter away the isoforms with comparatively fewer count number}
}
\value{
a dataframe
}
\description{
Function to map proteins to known protein complex database
}
\examples{
\dontrun{
  MOLM <- imprints_complex_mapping(MOLM, MOLM_score, treatment="TNF", targetcategory=c("NC","CC"))
  MOLM <- imprints_complex_mapping(MOLM, MOLM_category, treatment="TNF", complexID=c(1,10,1234))
}


}
