% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_corr_in_complex.R
\name{imprints_corr_in_complex}
\alias{imprints_corr_in_complex}
\title{imprints_corr_in_complex}
\usage{
imprints_corr_in_complex(
  data = NULL,
  nread = 6,
  goodcorrcutoff = 0.5,
  vgoodcorrcutoff = 0.9,
  removeredundancy = FALSE,
  similaritythreshold = 1,
  complexID = NULL,
  doplotting = TRUE,
  layout = c(3, 2),
  external = TRUE,
  toplabel = ""
)
}
\arguments{
\item{data}{dataset after imprints_complex_mapping()}

\item{nread}{number of reading channels or sample treatments, default value 6}

\item{goodcorrcutoff}{the threshold for a good correlation, default value is 0.5}

\item{vgoodcorrcutoff}{the threshold for a very good correlation, default value is 0.9}

\item{removeredundancy}{scrutinize the complex input, to remove the redundancy
according to the measured subunits, default set to FALSE}

\item{similaritythreshold}{the threshold for similarity, numeric number between 0 to 1,
default use 1.0 to only remove the complete overlap ones}

\item{complexID}{a vector of complexID to retrieve from the data, if specified}

\item{doplotting}{whether to generate the correlation plot, default set to TRUE}

\item{layout}{a vector indicating the panel layout for multi-panel plots per page}
}
\value{
a dataframe
}
\description{
Function to calculate the profile correlation between the subunit proteins of
the protein complexes found in IMPRINTS dataset
}
\examples{
\dontrun{
  MOLM <- imprints_corr_in_complex(MOLM, nread=6)
}


}
