% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_correlation.R
\name{imprints_correlation}
\alias{imprints_correlation}
\title{imprints_correlation}
\usage{
imprints_correlation(
  data = NULL,
  setvector = NULL,
  treatmentvector = NULL,
  bygene = TRUE
)
}
\arguments{
\item{data}{dataset of average IMPRINTS profile, that is after calculating the relative protein
abundance differences and deriving the average, i.e., after imprints_average(), make sure
the columns with readings are named in the format like "37C_Treatment"}

\item{setvector}{if necessary, a two-element character vector with distinguishable keyword
from the column names of the input dataset}

\item{treatmentvector}{a two-element character vector with distinguishable keyword
from the column names of the input dataset}

\item{bygene}{whether to match the protein accessions by gene, default set to TRUE,
otherwise would match be protein accession which could be affected by protein isoforms,
when there is isoform difference, the first dataset would be used as the reference name}
}
\value{
a dataframe
}
\description{
Function to calculate the Pearson correlation of two treatment conditions
}
\examples{
\dontrun{
  MOLM <- imprints_correlation(MOLM, setvector=NULL, treatmentvector=c("T1","T2"))
  MOLM <- imprints_correlation(MOLM, setvector=c("M13","M16"), treatmentvector=c("T1","T2"))
}


}
