% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_diffExp.R
\name{imprints_diffExp}
\alias{imprints_diffExp}
\title{imprints_diffExp}
\usage{
imprints_diffExp(
  data,
  set = NULL,
  basetemp = "37C",
  contrast = NULL,
  logFC_threshold = 0.3,
  adjp_threshold = 0.01,
  pfdatabase = FALSE,
  labelnodes = TRUE,
  xlimit = NULL,
  ylimit = NULL,
  labelgeneid = NULL,
  returneset = FALSE
)
}
\arguments{
\item{data}{dataset after imprints_normalization() function, readings in log2 format}

\item{set}{a single character to indicate the sample name}

\item{basetemp}{the character indicating the baseline temperature for
expression levels, default value is 37C}

\item{contrast}{a character to indicate the contrasting treatment conditions}

\item{logFC_threshold}{the threshold value for log fold changes, default set at 0.3}

\item{adjp_threshold}{the threshold value for adjusted p values, default set at 0.01}

\item{labelnodes}{whether to label the proteins with significant differential expression}

\item{xlimit}{a two numeric element vector to indicate the limit of x-axis}

\item{ylimit}{a two numeric element vector to indicate the limit of y-axis}

\item{labelgeneid}{a vector of the gene symbol id to show on the plot}
}
\value{
a dataframe
}
\description{
Function to have an overview of the IMPRINTS-CETSA dataset, regarding to
expression level changes only, to select out the differentially expressed proteins
}
\examples{
\dontrun{
  MOLM <- imprints_diffExp(MOLM, set="M13", contrast="TNFa-DMSO")
}

}
