% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_globalview.R
\name{imprints_globalview}
\alias{imprints_globalview}
\title{imprints_globalview}
\usage{
imprints_globalview(
  data,
  basetemp = "37C",
  qc = TRUE,
  cvthreshold = 0.1,
  corrthreshold = 0.5,
  abundancechange_cutoff = NULL,
  stabilitychange_cutoff = NULL,
  abundancechange_nMAD = 2.5,
  stabilitychange_nMAD = 2.5,
  labelnodes = FALSE,
  labelcategory = c("CC", "NC", "CN"),
  labelgeneid = NULL,
  xrange = NULL,
  yrange = NULL
)
}
\arguments{
\item{data}{dataset after imprints_caldiff() function}

\item{basetemp}{the character indicating the baseline temperature for
abundance/expression levels, default value is 37C}

\item{qc}{whether to apply the reproducibility quality control on input data,
default set to TRUE}

\item{cvthreshold}{the CV threshold value for subsetting reproducible
measurements, default value is 0.1}

\item{corrthreshold}{the Correlation threshold value for subsetting
reproducible measurements, default value is 0.5}

\item{abundancechange_nMAD}{the number of MADs to set the significance
cutoff on protein abundance level change, default value is 2.5}

\item{stabilitychange_nMAD}{the number of MADs to set the significance cutoff
on protein thermal stability shift, default value is 2.5}

\item{labelnodes}{whether to text-label the selected nodes, default set to FALSE}

\item{labelcategory}{the categories of nodes to label, default value is c("CC","NC","CN")}

\item{labelgeneid}{a vector of the gene symbol id to show on the plot, exclusive from labelcategory}
}
\value{
a dataframe
}
\description{
Function to have a global overview of the IMPRINTS-CETSA dataset, regarding to
protein abundance level changes and stability shifts
}
\examples{
\dontrun{
  MOLM <- imprints_globalview(MOLM)
}

}
