% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_imputation.R
\name{imprints_imputation}
\alias{imprints_imputation}
\title{imprints_imputation}
\usage{
imprints_imputation(
  data,
  iteration = 3,
  method = "norm.predict",
  percondition = FALSE,
  excludeNA = TRUE
)
}
\arguments{
\item{data}{Normalized dataset but with missing values to be imputed and filled in}

\item{iteration}{an integer specifying how many rounds of imputation should be performed}

\item{percondition}{whether do the imputation on each condition separately, default
set to FALSE, to be implemented}
}
\value{
a dataframe
}
\description{
Function to impute the missing values in IMPRINTS-CETSA dataset.
}
\examples{
\dontrun{
 MOLM <- imprints_imputation(MOLM)
}


}
