% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_normalization.R
\name{imprints_normalization}
\alias{imprints_normalization}
\title{imprints_normalization}
\usage{
imprints_normalization(datafile, isfile = FALSE, todf = TRUE, pertemp = TRUE)
}
\arguments{
\item{datafile}{a directory pathway pointing to a file with pre-normalization
data and experimental scheme information, under the following keywords:
set, replicate, temperature, treatment}

\item{isfile}{whether the input datafile is a file, which should contain the
annotation rows on the top of each quantitative column of the saved dataset
with keyword information of “temperature”, “treatment”, “replicate”, or
“set” in the case of complex experimental layout, default set to FALSE}

\item{todf}{whether to return the normalized data in a dataframe format,
default set to TRUE}

\item{pertemp}{whether do the normalization on each heating temperature separately,
default set to TRUE}
}
\value{
a dataframe
}
\description{
Function to perform a systematic normalization on the whole dataset
First IMPRINTS-CETSA data is transformed into an eSet format, then a VSN-based model normalization is applied,
and finally the function ‘removeBatchEffect’ from limma R package is called to remove potential batch effect.
}
\examples{
\dontrun{
 MOLM <- imprints_normalization(MOLM)
}


}
