% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_rawread.R
\name{imprints_rawread}
\alias{imprints_rawread}
\title{imprints_rawread}
\usage{
imprints_rawread(
  filevector,
  fchoose = FALSE,
  treatment = NULL,
  nread = 10,
  abdread = TRUE,
  noratiodata = TRUE,
  PDversion = 21,
  fdrcontrol = FALSE,
  refchannel = "126",
  channels = NULL
)
}
\arguments{
\item{filevector}{a file name or a vector of filenames to import}

\item{fchoose}{whether to choose file interactively, default set to FALSE}

\item{treatment}{a vector of treatment names applied to CETSA samples, in the same order as channels,
in a typical IMPRINTS/2D working scheme, the treatment name should contain both replicate and treatment
information, preferably concatenated with an underline}

\item{nread}{number of reading channels, should match the number of channels used, default value 10}

\item{abdread}{whether to read in protein abundance data, default set to TRUE}

\item{noratiodata}{whether to remove the protein ratio data, default set to TRUE}

\item{PDversion}{which version of Proteome Discoverer the data is searched, possible values 20,21,22,24}

\item{fdrcontrol}{whether to check the protein FDR confidence level, default set to FALSE}

\item{refchannel}{names of reference channel used in Proteome Discoverer search, default value 126}

\item{channels}{names of the read-in channels, default value NULL, it would automatically
match the provided channel number when it is 10 or 11}
}
\value{
a dataframe
}
\description{
Function to parse and read in IMPRINTS-CETSA data from tab delimited files exported from Proteome Discoverer
}
\examples{
\dontrun{
 MOLM13 <- imprints_rawread(c("M13_37C_frac_Proteins.txt","M13_52C_frac_Proteins.txt","M13_58C_frac_Proteins.txt"), nread=10,
 treatment=("B1_DMSO","B1_TNFa","B1_AT26533","B2_DMSO","B2_TNFa","B2_AT26533","B3_DMSO","B3_TNFa","B3_AT26533","Mix"))
}



}
