% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_rearrange.R
\name{imprints_rearrange}
\alias{imprints_rearrange}
\title{imprints_rearrange}
\usage{
imprints_rearrange(
  data,
  nread = 9,
  repthreshold = 0.75,
  withabdreading = TRUE,
  basetemp = "37C",
  averagecount = TRUE,
  countthreshold = 2,
  extraid = NULL
)
}
\arguments{
\item{data}{dataset to transform into a wide format}

\item{nread}{number of reading channels, default value 9 in typical IMPRINTS-CETSA scheme}

\item{repthreshold}{the minimal percentage threshold of protein being sampled
from multiple runs, default value is 0.75}

\item{withabdreading}{whether the kept proteins should have readings at abundance
reference channel, usually 37C or the lowest heating temperature (basetemp)}

\item{basetemp}{the character indicating the baseline temperature for
expression levels, default value is 37C}

\item{averagecount}{whether to take the median of the abundance count numbers across
the measured temperature range and then use this value for filtering, default set to TRUE,
otherwise, filter the proteins according to the associated count numbers at each temperature}

\item{countthreshold}{the minimal threshold number of associated abundance
count of proteins, default value is 2}

\item{extraid}{a vector of UniprotID to be included in the subset dataset}
}
\value{
a dataframe
}
\description{
Function to rearrange IMPRINTS-CETSA data into a wide one-unique-protein-per-row format
}
\examples{
\dontrun{
 MOLM <- imprints_rearrange(MOLM, nread=9, countthreshold=2)
}


}
