% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_reproducible.R
\name{imprints_reproducible}
\alias{imprints_reproducible}
\title{imprints_reproducible}
\usage{
imprints_reproducible(
  data,
  set = NULL,
  treatment = NULL,
  cvthreshold = 0.1,
  corrthreshold = 0.5,
  returnlist = FALSE
)
}
\arguments{
\item{data}{dataset after imprints_caldiff() function}

\item{set}{a single character to indicate the sample name}

\item{treatment}{a single character to indicate the sample name, when not specified,
all the treatments in the dataset would be screened through}

\item{cvthreshold}{the CV threshold value for subsetting reproducible
measurements, default value is 0.1}

\item{corrthreshold}{the Correlation threshold value for subsetting
reproducible measurements, default value is 0.5}

\item{returnlist}{whether to return as list but not as dataframe,
default set to FALSE}
}
\value{
a list of dataframe, for each condition
}
\description{
Function to retrieve a reproducible subset of the IMPRINTS-CETSA dataset,
regarding to both expression level changes and thermal shifts.
This is different from imprints_reproducible_abundance() function,
which is performed on the dataset with protein abundance but not the logratio data.
}
\examples{
\dontrun{
  MOLM2 <- imprints_reproducible(MOLM1)
}

}
