% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_reproducible_abundance.R
\name{imprints_reproducible_abundance}
\alias{imprints_reproducible_abundance}
\title{imprints_reproducible_abundance}
\usage{
imprints_reproducible_abundance(
  data,
  set = NULL,
  treatment = NULL,
  cvthreshold = 0.1,
  returnlist = FALSE
)
}
\arguments{
\item{data}{dataset after imprints_normalization() function}

\item{set}{a single character to indicate the sample name}

\item{treatment}{a single character to indicate the sample name}

\item{cvthreshold}{the CV threshold value for subsetting reproducible
measurements, default value is 0.1}

\item{returnlist}{whether to return as list but not as dataframe,
default set to FALSE}
}
\value{
a list of dataframe, for each condition
}
\description{
Function to retrieve a reproducible subset of the IMPRINTS-CETSA dataset,
regarding to both expression level changes and thermal shifts.
This is different from imprints_reproducible() function,
which is performed on the logratio data but not normalized dataset with protein abundance.
}
\examples{
\dontrun{
  MOLM2 <- imprints_reproducible_abundance(MOLM1)
}

}
