% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_score.R
\name{imprints_score}
\alias{imprints_score}
\title{imprints_score}
\usage{
imprints_score(
  data,
  format = "9",
  basetemp = "37C",
  basecvcheck = FALSE,
  stabilitycvcheck = FALSE,
  pfdatabase = FALSE,
  cvcutoffthreshold = 2.5,
  allzscore = TRUE,
  weightbycv = TRUE,
  fdrthreshold = 0.01,
  useMAD = FALSE,
  nMAD = 2.5,
  cutoffvector = NULL,
  labelnodes = FALSE,
  labelcategory = c("CC", "NC", "CN"),
  labelgeneid = NULL
)
}
\arguments{
\item{data}{dataset after calculating the relative protein abundance differences
i.e., imprints_caldiff()}

\item{format}{choose between 9 and 4, indicating how many categories to segregate,
default value is 9, in which proteins are divided into 9 categories: NN, CN+, CN-,
NC+, NC-, CC++, CC+-, CC-+, CC--. When switch to 4, proteins are instead divided into
4 categories: NN, CN, NC, CC. The sign of + or - after N or C is determined by the
sign of abundance.score and stability.score.mean, respectively}

\item{basetemp}{the character indicating the baseline temperature for
expression levels, default value is 37C}

\item{basecvcheck}{whether to perform a quality check on the reproducibility of baseline
protein expression levels across the replicates, default set to FALSE}

\item{stabilitycvcheck}{whether to perform a quality check on the reproducibility of protein
stability trend across the replicates, default set to FALSE}

\item{cvcutoffthreshold}{the significance level of threshold used for CV quality control,
default value is 2.5, typically can range from 2 to 3}

\item{allzscore}{when calculating the z score, whether to use all the readings from
all the different treatment groups, default set to TRUE}

\item{weightbycv}{whether to weight the thermal shift with the cv information, default set to TRUE}

\item{fdrthreshold}{the significance level of global fdr, default value is 0.01}

\item{useMAD}{whether to use MAD scheme for significance analysis, default set to FALSE}

\item{nMAD}{the significance level of MAD, default value 2.5}

\item{cutoffvector}{a vector to specify the symmetrical cutoff values (abundance, stability)}

\item{labelnodes}{whether to text-label the selected nodes, default set to FALSE}

\item{labelcategory}{the categories of nodes to label, default value is c("CC","NC","CN")}

\item{labelgeneid}{a vector of the gene symbol id to show on the plot, exclusive from labelcategory}
}
\value{
a dataframe
}
\description{
Function to calculate the relative abundance score and thermal stability score
Based on the scores, it divides proteins into 9 or 4 categories and shows in a scatter plot
}
\examples{
\dontrun{
  MOLM <- imprints_score(MOLM)
}


}
