% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_score_boundness.R
\name{imprints_score_boundness}
\alias{imprints_score_boundness}
\title{imprints_score_boundness}
\usage{
imprints_score_boundness(
  data,
  format = "9",
  basetemp = "37C",
  meancutoff = 0.25,
  boundedness = 4,
  qualitycutoff = 0.15,
  meandev = 0.15
)
}
\arguments{
\item{data}{dataset after calculating the relative protein abundance differences
i.e., imprints_caldiff()}

\item{format}{choose between 9 and 4, indicating how many categories to segregate,
default value is 9, in which proteins are divided into 9 categories: NN, CN+, CN-,
NC+, NC-, CC++, CC+-, CC-+, CC--. When switch to 4, proteins are instead divided into
4 categories: NN, CN, NC, CC. The sign of + or - after N or C is determined by the
sign of abundance.score and stability.score.mean, respectively}

\item{basetemp}{character indicating the baseline temperature for expression levels, default value is 37C}

\item{meancutoff}{The threshold value for the mean of triplicate measurement (First filter of hits)}

\item{boundedness}{Number of SEM, used for setting a bound that the mean needs to be within}

\item{qualitycutoff}{The threshold value for the SEM to separate well-measured and ill-measured measurements}

\item{meandev}{The deviation of individual mean from the group mean in CN categorization}
}
\value{
a list
}
\description{
Function to categorize protein according to their Expression/Stability change
}
\examples{
\dontrun{
  MOLM <- imprints_score_boundness(MOLM)
}


}
