% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_plots_arrange.R
\name{ms_plots_arrange}
\alias{ms_plots_arrange}
\title{ms_plots_arrange}
\usage{
ms_plots_arrange(
  plotlist,
  innerlayout = NULL,
  mainlayout = c(4, 1),
  external = TRUE,
  toplabel = "",
  leftlabel = "",
  bottomlabel = "",
  pdfname = "mixed_ggplotting.pdf",
  pdfheight = 12,
  pdfwidth = 10,
  returnplots = FALSE
)
}
\arguments{
\item{plotlist}{a list of plots objects (now hardcode for the case of 2,3,4)}

\item{innerlayout}{a vector indicating the lower level panel layout for
each multi-panel plot, default value is NULL, would change according to
the number of supplied plots}

\item{mainlayout}{a vector indicating the main panel layout for multi-panel
plots per page, default value is c(4,1)}

\item{toplabel}{textual label at the top part of the page}

\item{leftlabel}{textual label at the left side of the page}

\item{bottomlabel}{textual label at the bottom part of the page}

\item{pdfheight}{a number indicate the height of pdf file, default value 12}

\item{pdfwidth}{a number indicate the width of pdf file, default value 12}
}
\value{
a list of ggplot2 object
}
\description{
Function to generate pdf files with multipanel compiled from plots objects
}
\examples{
\dontrun{
  ms_plots_arrange(MOLM)
}


}
\keyword{internal}
